requirements.txt
beautifulsoup4==4.12.2
matplotlib==3.7.4
pandas==2.0.3
python_dateutil==2.8.2
requests==2.28.1
streamlit==1.28.2
lxml==4.9.3